bundle epub and html modules in dialect module
apart from being a much more plausible structure (previously,
both had dialect submodules, now both are submodules of a single
dialects module), this is also less error-prone. Since there exists
an html module in the python standard library, there was a possibility
for confusion about which module gets imported, based on the import
order of the specific environment. Not exposing `html` as an toplevel
module gets rid of this possibility.